feat: finalizando envio aos endpoints

+++ src/main/java/com/example/demo/repository/FilmeRepository.java
@@ -3,4 +3,5 @@ package com.example.demo.repository;
 import com.example.demo.model.Filme;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FilmeRepository extends JpaRepository<Filme, Integer> {}
+public interface FilmeRepository extends JpaRepository<Filme, Long> { }
+
